+++ crate/ir_model/src/node.rs
@@ -1,6 +1,6 @@
 pub use self::{
     node_copy_text::NodeCopyText, node_hierarchy::NodeHierarchy, node_id::NodeId,
-    node_inbuilt::NodeInbuilt, node_names::NodeNames,
+    node_inbuilt::NodeInbuilt, node_names::NodeNames, node_ordering::NodeOrdering,
 };
 
 mod node_copy_text;
@@ -8,3 +8,4 @@ mod node_hierarchy;
 mod node_id;
 mod node_inbuilt;
 mod node_names;
+mod node_ordering;