+++ crate/taffy_model/src/lib.rs
@@ -13,6 +13,11 @@ pub use crate::api_doc::ApiDoc;
 #[cfg(all(feature = "openapi", not(feature = "test")))]
 mod api_doc;
 
+/// Default text font size.
+pub const TEXT_FONT_SIZE: f32 = 11.0f32;
+/// Default text line height.
+pub const TEXT_LINE_HEIGHT: f32 = 13.0f32;
+
 pub use crate::{
     diagram_lod::DiagramLod, dimension::Dimension, dimension_and_lod::DimensionAndLod,
     entity_highlighted_span::EntityHighlightedSpan,